# docs/openapi/mdm-openapi1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MDM Device Posture API",
    "version": "1.1.0",
    "description": "Canonical API contract for MDM  device posture system. This JSON spec covers agent ingestion, normalization, evaluation, scoring, decisioning, remediation, policy management, catalog, lookups, and admin jobs."
  },
  "servers": [
    {
      "url": "https://mdm.24onlinebilling.com",
      "description": "Production"
    },
    {
      "url": "https://staging-mdm.24onlinebilling.com",
      "description": "Staging"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/agent/posture-payloads": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "Submit posture payload",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosturePayloadIngestRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosturePayloadIngestResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "List posture payloads",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          },
          {
            "in": "query",
            "name": "device_external_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "process_status",
            "schema": {
              "$ref": "#/components/schemas/PayloadProcessStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payload list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/agent/posture-payloads/{payload_id}": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Get posture payload",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/PayloadId"
          }
        ],
        "responses": {
          "200": {
            "description": "Payload detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicePosturePayload"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/v1/agent/devices/{device_external_id}/decision/latest": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Get latest decision for a device",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/DeviceExternalId"
          }
        ],
        "responses": {
          "200": {
            "description": "Latest decision response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDecisionResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/agent/decision-responses/{response_id}/ack": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "Acknowledge decision response",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/ResponseId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionAckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionAckResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "XTenantId": {
        "in": "header",
        "name": "X-Tenant-Id",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "Page": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
      "Size": {
        "in": "query",
        "name": "size",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 20
        }
      },
      "Sort": {
        "in": "query",
        "name": "sort",
        "schema": {
          "type": "string",
          "example": "created_at,desc"
        }
      },
      "PayloadId": {
        "in": "path",
        "name": "payload_id",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "DeviceExternalId": {
        "in": "path",
        "name": "device_external_id",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "ResponseId": {
        "in": "path",
        "name": "response_id",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          }
        }
      },
...
      "PagedEnvelope": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "page": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "size": {
                "type": "integer"
              },
              "total_elements": {
                "type": "integer"
              },
              "total_pages": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "ComplianceAction": {
        "type": "string",
        "enum": [
          "ALLOW",
          "NOTIFY",
          "QUARANTINE",
          "BLOCK"
        ]
      },
...
      "PayloadProcessStatus": {
        "type": "string",
        "enum": [
          "RECEIVED",
          "VALIDATED",
          "EVALUATED",
          "FAILED"
        ]
      },
...
      "DeliveryStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "SENT",
          "ACKED",
          "FAILED",
          "TIMEOUT"
        ]
      },
...
      "PosturePayloadIngestRequest": {
        "type": "object",
        "required": [
          "device_external_id",
          "agent_id",
          "payload_version",
          "payload_json"
        ],
        "properties": {
          "device_external_id": {
            "type": "string"
          },
          "agent_id": {
            "type": "string"
          },
          "payload_version": {
            "type": "string"
          },
          "payload_hash": {
            "type": "string"
          },
          "capture_time": {
            "type": "string",
            "format": "date-time"
          },
          "payload_json": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "PosturePayloadIngestResponse": {
        "type": "object",
        "properties": {
          "payload_id": {
            "type": "integer",
            "format": "int64"
          },
          "process_status": {
            "$ref": "#/components/schemas/PayloadProcessStatus"
          },
          "evaluation_run_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        }
      },
...
      "DevicePosturePayload": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tenant_id": {
            "type": "string",
            "nullable": true
          },
          "device_external_id": {
            "type": "string"
          },
          "agent_id": {
            "type": "string",
            "nullable": true
          },
          "payload_version": {
            "type": "string",
            "nullable": true
          },
          "payload_hash": {
            "type": "string",
            "nullable": true
          },
          "process_status": {
            "$ref": "#/components/schemas/PayloadProcessStatus"
          },
          "payload_json": {
            "type": "object",
            "additionalProperties": true
          },
          "received_at": {
            "type": "string",
            "format": "date-time"
          },
          "processed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "DeviceDecisionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "posture_evaluation_run_id": {
            "type": "integer",
            "format": "int64"
          },
          "tenant_id": {
            "type": "string",
            "nullable": true
          },
          "device_external_id": {
            "type": "string"
          },
          "decision_action": {
            "$ref": "#/components/schemas/ComplianceAction"
          },
          "trust_score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "remediation_required": {
            "type": "boolean"
          },
          "delivery_status": {
            "$ref": "#/components/schemas/DeliveryStatus"
          },
          "response_payload": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "DecisionAckRequest": {
        "type": "object",
        "required": [
          "delivery_status"
        ],
        "properties": {
          "delivery_status": {
            "type": "string",
            "enum": [
              "ACKED",
              "FAILED",
              "TIMEOUT"
            ]
          },
          "acknowledged_at": {
            "type": "string",
            "format": "date-time"
          },
          "error_message": {
            "type": "string"
          }
        }
      },
      "DecisionAckResponse": {
        "type": "object",
        "properties": {
          "response_id": {
            "type": "integer",
            "format": "int64"
          },
          "delivery_status": {
            "$ref": "#/components/schemas/DeliveryStatus"
          },
          "acknowledged_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  }
}